fixed warning generating code in GBProcedure.h; fixed code, which could not be compiled in multivariatepolynomial::substitute; type for local ids in termadditionmanager is unsigned instead of short, which was not enough for smtrat on some non-linear instances

+++ src/carl/util/TermAdditionManager.h
@@ -22,7 +22,7 @@ namespace carl
 template<typename Polynomial>
 class TermAdditionManager {
 public:
-	typedef unsigned short IDType;
+	typedef unsigned IDType;
 	typedef typename Polynomial::CoeffType Coeff;
 	typedef Term<Coeff> TermType;
 	typedef TermType TermPtr;
